Any name with {time}[smhdw] as second arg does the ControlPersist trick
Dropped the use of 0.0/8 as trigger for creating ssh persistent
connection socket; it was used since some TCP/IP stacks EINVAL's
connection attempt to such an address and failing fast was
convenient in a case persistent connection socket was not live.

But now, some TCP/IP stacks don't fail early but try to connect
to the 0.* address -- waiting it to fail (or worse, connect!).

This behavioral difference can be noticed between Linux kernels
5.2 and 5.3 -- 5.2 EINVALs, 5.3 tries to connect. See commit
9917d541c8c3 in Linus' git repo for details (and this commit diff).

Therefore it is better give user "responsibility" to choose
whatever "name" they want, valid or invalid address which
works for them, something (open)ssh does not allow is at
least not a name help is triggered...